classflow/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
django-extensions==3.2.3
django-formtools==2.5.1
django-otp==1.4.1
django-phonenumber-field==7.3.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-two-factor-auth==1.16.0
gunicorn==22.0.0
idna==3.7
MarkupSafe==2.1.5
packaging==24.0
phonenumbers==8.13.34
pillow==10.3.0
pycparser==2.22
pyOpenSSL==24.1.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
recaptcha==1.0rc1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
